% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/junkgen.R
\name{junkgen}
\alias{junkgen}
\title{Generate a single pseudo-random number using the MLS Junk Generator algorithm}
\usage{
junkgen(w, x, y, z)
}
\arguments{
\item{w}{the first seed required by the MLS Junk Generator algorithm}

\item{x}{the first seed required by the MLS Junk Generator algorithm}

\item{y}{the first seed required by the MLS Junk Generator algorithm}

\item{z}{the first seed required by the MLS Junk Generator algorithm}
}
\value{
A numeric vector containing a single pseudo-random number
}
\description{
Based on user input seeds, this function generates a pseudo-random number.
This is called by the mlsjunkgen package's other functions to generate a
pseudo-random number stream.
}
\examples{
# Generate a pseudo-random number with user-specified seeds

w <- 1
x <- 2
y <- 3
z <- 4
junkgen(w = w, x = x, y = y, z = z) # returns "[1] 0.9551644"
}

