% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlsjunkgen.R
\docType{package}
\name{mlsjunkgen}
\alias{mlsjunkgen}
\alias{mlsjunkgen-package}
\title{mlsjunkgen: Use the MLS Junk Generator Algorithm to Generate a Stream of Pseudo-Random Numbers}
\description{
mlsjunkgen: Use the MLS Junk Generator Algorithm to Generate a Stream of Pseudo-Random Numbers
}
\section{mlsjunkgen functions}{

\itemize{
 \item \code{\link{junkgen}}: generate a single pseudo-random number; called by the other functions
 \item \code{\link{mlsjunkgenv}}: generate a vector stream of pseudo-random numbers
 \item \code{\link{mlsjunkgend}}: generate a data frame of pseudo-random numbers
 \item \code{\link{mlsjunkgenm}}: generate a matrix of pseudo-random numbers
}
}

